♻️ Refactor: 프론트 폴더 구조(간단) 정리
Refactor: 프론트 폴더 구조(간단) 정리
