+++ src/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import LoginPage from './pages/LoginPage';
-import SignupPage from './pages/SignupPage';
+import LoginPage from './pages/login/LoginPage.js';
+import SignupPage from './pages/login/SignupPage.js';
 import MapPage from './pages/MapPage.tsx';
 import DistributionMapPage from './pages/DistributionMapPage.tsx';
 import PriceMapPage from './pages/PriceMapPage.tsx';
-import WholesalePricePage from './pages/Wholesale-price.tsx';
-import RetailPricePage from './pages/Retail-price.tsx';
+import WholesalePricePage from './pages/wholesale_retail/Wholesale-price.tsx';
+import RetailPricePage from './pages/wholesale_retail/Retail-price.tsx';
 
 function App() {
   return (
