+++ src/api/axiosInstance.js
@@ -5,7 +5,7 @@ const axiosInstance = axios.create({
   // withCredentials: true, // 필요 시 쿠키 인증 사용
 });
 
-// ✅ 요청 인터셉터 추가 – JWT 토큰 자동 삽입
+// ❗ 요청 인터셉터 추가 – JWT 토큰 자동 삽입 ❗ 
 axiosInstance.interceptors.request.use(
   (config) => {
     const token = localStorage.getItem("token");
