+++ src/pages/login/SignupPage.js
@@ -1,4 +1,4 @@
-import SignupForm from "../components/SignupForm";
+import SignupForm from "../../components/login/SignupForm";
 function SignupPage() {
   return (
     <div>
